Ensure the input file exists and fix exit codes
The input file is checked for existence and also set in the main function
(as opposed to the parse_args function). It is also now a non-constant
FILE pointer, which will avoid having to open the file every time it needs
to be accessed without having a shared reference.

Exit codes where a fatal error has occurred and the program has not
completed the process that it was supposed to are now EXIT_FAILURE, as
opposed to EXIT_SUCCESS. This will allow the program to better integrate
with bash scripts and the like.