Add help option and begin implementing multi-file support
The parse_args function now returns an integer that corresponds to the
index of the first file in `argv`. The input file arguments are placed at
the end of array by the `getopt_long()` function. Similarly, the global
variable `input_file` has been removed, as there is no longer a need for
it.

Currently, the program lacks functionality in dealing with the input
files, and so they are simply printed out before the program exits.

A help option has been added, as per common practice in many *nix
utilities and programs.